feat(reparto): ship confirmation and liquidation stack
Additive JAVIER DDL 036/037 already applied; keep finance/confirm
capabilities fail-closed until phased enable. DSEDAC untouched.

Refs gmp_app_mobilidad-xlx

+++ lib/features/repartidor/presentation/widgets/rutero_detail_finalize.dart
@@ -31,7 +31,7 @@ class RuteroDetailFinalize extends StatelessWidget {
 
   final AlbaranEntrega albaran;
   final List<EntregaItem> items;
-  final Map<String, int> productQuantities;
+  final Map<String, double> productQuantities;
   final TextEditingController nombreController;
   final TextEditingController dniController;
   final TextEditingController observacionesController;
